stage1_Q2#VQE/requirements.txt
pennylane
pyzx
cqlib
isqopen
